Версия 0.0.5. Исправлена логика методов, убраны очевидные ошибки, скорректировано разделение пользователей.

+++ MVP_TaskManager/Services/Operations_users.cs
@@ -3,21 +3,27 @@ using MVP_TaskManager.Data;
 using MVP_TaskManager.DTO;
 using MVP_TaskManager.Models;
 using AsyncTask = System.Threading.Tasks;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
 
 namespace MVP_TaskManager.Classes
 {
     public class Operations_users
     {
         private readonly TaskManagerContext context;
-
-
         public Operations_users(TaskManagerContext _context)
         {
             context = _context;
         }
 
-        public async Task<User> CreateNewUser(UserDTO user)
+        public async Task<User> CreateNewUser(User_CreateDTO user)
         {
+            
+            if (Enum.IsDefined(typeof(UserRole), "Admin") == false)
+            {
+                throw new InvalidOperationException("Такой роли нет!");
+            }
 
             var newUser = new User
             {
@@ -25,26 +31,50 @@ namespace MVP_TaskManager.Classes
                 Login = user.Login,
                 Password = user.Password,
                 RegDate = DateOnly.FromDateTime(DateTime.UtcNow),
+                Role = UserRole.User,
             };
+
+            if (await CheckLogin(newUser.Login))
+            {
+                throw new InvalidOperationException("Такой логин уже существует!");
+            }
 
             context.Users.Add(newUser);
             await context.SaveChangesAsync();
             return newUser;
         }
 
-        public async Task<bool> UpdateUser(int id_user, UserDTO user)
+
+
+        public async Task<bool> UpdateUser(int id_user, 
+            [FromQuery] User_UpdateDTO user, 
+            int id_User_Updating, 
+            bool is_Admin)
         {
             var userForUpdate = await context.Users
                 .FindAsync(id_user);
 
             if (userForUpdate == null) return false;
 
-            userForUpdate.Username = user.Username;
-            userForUpdate.Password = user.Password;
-            userForUpdate.RegDate = user.RegDate;
+            if (user.Username != null)
+                userForUpdate.Username = user.Username;
+
+            if (user.Password != null)
+                userForUpdate.Password = user.Password;
+
+            if (CheckRoots(id_user, is_Admin, id_User_Updating) == false)
+            {
+                throw new InvalidOperationException
+                    ("Недостаточно прав на эту операцию!");
+            }
 
             await context.SaveChangesAsync();
             return true;
+        }
+
+        private async Task<bool> CheckLogin(string login)
+        {
+            return await context.Users.AnyAsync(u => u.Login == login);
         }
 
         public async Task<List<User>> AllUser()
@@ -58,14 +88,49 @@ namespace MVP_TaskManager.Classes
             return userByID;
         }
 
-        public async Task<bool> DeleteUser(int id_user)
+        public async Task<bool> DeleteUser(int id_user, bool is_Admin, int id_User_Deleting )
         {
+            
             var userForDel = await context.Users
-                .FindAsync(id_user);
+                .FirstOrDefaultAsync(u => u.Id == id_user);
 
             if (userForDel == null) return false;
 
+            Console.WriteLine($"id_user = {id_user}");
+            Console.WriteLine($"id_User_Deleting = {id_User_Deleting}");
+            Console.WriteLine($"is_Admin = {is_Admin}");
+
+            if (CheckRoots(id_user, is_Admin, id_User_Deleting) == false)
+            {
+                throw new UnauthorizedAccessException();
+            }
+
+            await DeleteAllTasksByUser(id_user);
             context.Users.Remove(userForDel);
+            
+            await context.SaveChangesAsync();
+            return true;
+        }
+
+        public bool CheckRoots(int? id_user, bool is_Admin, int id_User_Required)
+        {
+            if (id_user == id_User_Required || is_Admin)
+            {
+                return true;
+            }
+            else return false;
+        }
+
+        private async Task<bool> DeleteAllTasksByUser (int id_user)
+        {
+            var tasks = await context.Tasks
+                .Where(t => t.IdUser == id_user)
+                .ToListAsync();
+
+            if(tasks.Count == 0)
+                return false;
+
+            context.Tasks.RemoveRange(tasks);
             await context.SaveChangesAsync();
             return true;
         }
